Provider: Redirect errors with state (#140)
* override_settings

* fix: handled exception in case of mising ta

* fix:  improved url definition

* tests:  test validate_entity_metadata

* fix: changed status response when missing trust anchor

* tests: reverse in test

* fix: MAX_ENTRIES_PAGE

* fix:  User claims in CIE idToken

* tests: test claims cie id_token

* fix: redirect errors with state

* fix: self.payload.get("state", "")

* fix: started to divide provider views in different file

* fix: divided views in different file

* fix: self.payload.get("state", "")

* fix: spid_cie_oidc.provider.views divided in different files
